# CTG_monitor/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ActiveSessionsResponse:
    description: Список всех активных сессий мониторинга
    properties:
      count:
        description: Количество активных сессий
        example: 3
        type: integer
      sessions:
        description: Список активных сессий
        items:
          $ref: '#/definitions/handlers.SessionResponse'
        type: array
    type: object
  handlers.CardSessionsResponse:
    description: Список сессий для конкретной медицинской карты
    properties:
      card_id:
        description: UUID медицинской карты
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      count:
        description: Количество сессий
        example: 5
        type: integer
      sessions:
        description: Список сессий
        items:
          $ref: '#/definitions/handlers.SessionResponse'
        type: array
    type: object
  handlers.CleanupResponse:
    description: Результат операции очистки зависших сессий
    properties:
      active_sessions:
        description: Количество активных сессий после очистки
        example: 2
        type: integer
      message:
        description: Сообщение о результате
        example: Очистка сессий выполнена
        type: string
    type: object
  handlers.DeviceStatusResponse:
    description: Текущий статус устройства КТГ
    properties:
      device_id:
        description: Идентификатор устройства
        example: CTG-DEVICE-001
        type: string
      duration:
        description: Продолжительность активной сессии в секундах
        example: 3600
        type: integer
      session_id:
        description: UUID активной сессии (если есть)
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      start_time:
        description: Время начала активной сессии
        example: "2023-09-01T10:00:00Z"
        type: string
      status:
        description: Статус устройства
        enum:
        - active
        - idle
        example: active
        type: string
    type: object
  handlers.DevicesResponse:
    description: Список всех доступных устройств КТГ
    properties:
      count:
        description: Количество устройств
        example: 2
        type: integer
      devices:
        description: Список идентификаторов устройств
        example:
        - CTG-DEVICE-001
        - CTG-DEVICE-002
        items:
          type: string
        type: array
    type: object
  handlers.ErrorResponse:
    description: Стандартная структура ответа об ошибке
    properties:
      details:
        description: Дополнительные детали ошибки
        example: field required
        type: string
      error:
        description: Описание ошибки
        example: Неверный формат данных
        type: string
    type: object
  handlers.HealthResponse:
    description: Информация о состоянии и работоспособности сервиса
    properties:
      active_sessions:
        description: Количество активных сессий
        example: 3
        type: integer
      service:
        description: Название сервиса
        example: CTG Monitor
        type: string
      status:
        description: Статус сервиса
        example: healthy
        type: string
      timestamp:
        description: Время проверки
        example: "2023-09-01T10:00:00Z"
        type: string
    type: object
  handlers.SessionDataResponse:
    description: Данные мониторинга КТГ, собранные во время сессии
    properties:
      fhr_data:
        description: Данные частоты сердечных сокращений плода
      session_id:
        description: UUID сессии
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      total_points:
        description: Общее количество точек данных
        example: 1250
        type: integer
      uc_data:
        description: Данные маточных сокращений
    type: object
  handlers.SessionRequest:
    description: Данные для создания новой сессии мониторинга
    properties:
      card_id:
        description: UUID медицинской карты пациента
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      device_id:
        description: Идентификатор устройства КТГ
        example: CTG-DEVICE-001
        type: string
    required:
    - card_id
    - device_id
    type: object
  handlers.SessionResponse:
    description: Информация о сессии мониторинга КТГ
    properties:
      card_id:
        description: UUID медицинской карты
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      device_id:
        description: Идентификатор устройства
        example: CTG-DEVICE-001
        type: string
      duration:
        description: Продолжительность в секундах
        example: 5400
        type: integer
      end_time:
        description: Время окончания сессии (если завершена)
        example: "2023-09-01T11:30:00Z"
        type: string
      session_id:
        description: UUID сессии
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      start_time:
        description: Время начала сессии
        example: "2023-09-01T10:00:00Z"
        type: string
      status:
        description: Статус сессии
        enum:
        - active
        - stopped
        example: active
        type: string
    type: object
  handlers.SuccessResponse:
    description: Стандартная структура успешного ответа
    properties:
      data:
        description: Дополнительные данные
      message:
        description: Сообщение об успехе
        example: Операция выполнена успешно
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support Team
    url: http://www.swagger.io/support
  description: API для системы мониторинга КТГ (кардиотокографии). Предоставляет возможности
    для управления сессиями мониторинга, работы с медицинскими картами и устройствами
    КТГ.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CTG Monitor API
  version: "1.0"
paths:
  /cards/{card_id}/sessions:
    get:
      description: Возвращает все сессии мониторинга КТГ для указанной медицинской
        карты
      parameters:
      - description: UUID медицинской карты
        format: uuid
        in: path
        name: card_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сессии медицинской карты
          schema:
            $ref: '#/definitions/handlers.CardSessionsResponse'
        "400":
          description: Неверный ID медицинской карты
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Ошибка получения сессий
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получение сессий для медицинской карты
      tags:
      - cards
  /devices:
    get:
      description: Возвращает список всех доступных устройств КТГ в системе
      produces:
      - application/json
      responses:
        "200":
          description: Список устройств
          schema:
            $ref: '#/definitions/handlers.DevicesResponse'
      summary: Получение списка устройств
      tags:
      - devices
  /devices/{device_id}/status:
    get:
      description: Возвращает текущий статус указанного устройства КТГ
      parameters:
      - description: Идентификатор устройства
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Статус устройства
          schema:
            $ref: '#/definitions/handlers.DeviceStatusResponse'
      summary: Получение статуса устройства
      tags:
      - devices
  /monitoring/cleanup:
    post:
      description: Выполняет очистку зависших и неактивных сессий в системе
      produces:
      - application/json
      responses:
        "200":
          description: Результат очистки
          schema:
            $ref: '#/definitions/handlers.CleanupResponse'
      summary: Очистка зависших сессий
      tags:
      - monitoring
  /monitoring/health:
    get:
      description: Возвращает информацию о текущем состоянии и работоспособности сервиса
        мониторинга КТГ
      produces:
      - application/json
      responses:
        "200":
          description: Состояние сервиса
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
      summary: Проверка состояния сервиса
      tags:
      - monitoring
  /sessions/{session_id}:
    get:
      description: Возвращает подробную информацию о конкретной сессии мониторинга
      parameters:
      - description: UUID сессии
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о сессии
          schema:
            $ref: '#/definitions/handlers.SessionResponse'
        "400":
          description: Неверный ID сессии
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Сессия не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получение информации о сессии
      tags:
      - sessions
  /sessions/{session_id}/data:
    get:
      description: Возвращает все данные мониторинга КТГ, собранные во время указанной
        сессии
      parameters:
      - description: UUID сессии
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные КТГ сессии
          schema:
            $ref: '#/definitions/handlers.SessionDataResponse'
        "400":
          description: Неверный ID сессии
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Сессия не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получение данных КТГ для сессии
      tags:
      - sessions
  /sessions/active:
    get:
      description: Возвращает список всех активных сессий мониторинга КТГ
      produces:
      - application/json
      responses:
        "200":
          description: Список активных сессий
          schema:
            $ref: '#/definitions/handlers.ActiveSessionsResponse'
      summary: Получение списка активных сессий
      tags:
      - sessions
  /sessions/start:
    post:
      consumes:
      - application/json
      description: Создает новую сессию мониторинга КТГ для указанной медицинской
        карты и устройства
      parameters:
      - description: Данные для создания сессии
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Сессия успешно запущена
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.SessionResponse'
              type: object
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Сессия для устройства уже активна
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Запуск новой сессии мониторинга КТГ
      tags:
      - sessions
  /sessions/stop/{session_id}:
    post:
      description: Завершает указанную активную сессию мониторинга КТГ
      parameters:
      - description: UUID сессии
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сессия успешно завершена
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.SessionResponse'
              type: object
        "400":
          description: Неверный ID сессии
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Сессия не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Завершение активной сессии мониторинга
      tags:
      - sessions
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API ключ для авторизации (если требуется)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
tags:
- description: Управление сессиями мониторинга КТГ. Позволяет создавать, завершать
    и получать информацию о сессиях мониторинга.
  name: sessions
- description: Работа с медицинскими картами пациентов. Получение истории сессий для
    конкретных пациентов.
  name: cards
- description: Управление устройствами КТГ. Мониторинг состояния и доступности медицинского
    оборудования.
  name: devices
- description: Мониторинг состояния сервиса. Проверка работоспособности системы и
    выполнение служебных операций.
  name: monitoring